IMP: Ported to Win64. FIX: Fixed 64bit size_t and unsigned int typo problems. FIX: Fixed many 64bit static analysis problem.
svn path=/trunk/; revision=758

+++ Source/ZEEngine/ZEGraphics/ZECanvas.h
@@ -142,7 +142,7 @@ class ZECanvas : public ZEDynamicVertexBuffer
 		bool							LoadFromFile(const ZEString& FileName);
 		void							SaveToFile(const ZEString& FileName);
 
-		virtual unsigned int			GetBufferSize();
+		virtual size_t					GetBufferSize();
 		virtual void*					GetVertexBuffer();
 		
 										ZECanvas();
